+++ public/js/ajax/auth/signup.ajax.js
@@ -20,35 +20,76 @@ $( document ).ready(function() {
             return;  // Stop further execution
         }
 
-        // Get CSRF token from meta tag
-        var csrfToken = $('meta[name="csrf-token"]').attr('content');
+        // This disables the button
+        $('#signupFormSubmit').attr('disabled', true)
 
-        $.ajax({
-            type: "POST",
-            url: "/backend/signup",
-            data: formData,
-            dataType: "json",
-            headers: {
-                'X-CSRF-TOKEN': csrfToken  // Add CSRF token to the request headers
-            },
-            success: function(response){
-                console.log(response);
-            },
-            error: function(xhr, status, error) {
-                console.log("Error occurred with the request:");
-                console.log("Status: " + status);
-                console.log("Error: " + error);
-                console.log("Response Text: " + xhr.responseText);  // Shows the response body from the server
-                try {
-                    var jsonResponse = JSON.parse(xhr.responseText);  // Try to parse the error response
-                    if (jsonResponse.errors) {
-                        console.log("Validation Errors:", jsonResponse.errors);
+        if (formData.password != formData.password_confirmation ){
+            Swal.fire({
+                title: 'Oops!',
+                text: 'Password does not match',
+                icon: 'error',
+                allowOutsideClick: false,
+                allowEscapeKey: false,
+                confirmButtonText: 'Okay',
+                confirmButtonColor: '#0054a6',
+            })
+            $('#signupFormSubmit').attr('disabled', false)
+        }
+        else {
+            // Get CSRF token from meta tag
+            var csrfToken = $('meta[name="csrf-token"]').attr('content');
+    
+            $.ajax({
+                type: "POST",
+                url: "/backend/signup",
+                data: formData,
+                dataType: "json",
+                headers: {
+                    'X-CSRF-TOKEN': csrfToken  // Add CSRF token to the request headers
+                },
+                success: function(response){
+                    Swal.fire({
+                        title: 'Success!',
+                        text: response.message,
+                        icon: 'success',
+                        allowOutsideClick: false,
+                        allowEscapeKey: false,
+                        confirmButtonText: 'Back to Log In',
+                        confirmButtonColor: '#0054a6',
+                    })
+                    .then((result) => {
+                        if (result.isConfirmed) {
+                            window.location.href = '/login';
+                        }
+                    })
+                    
+                },
+                error: function(response) {
+                    if (response.responseJSON.status == 'error'){
+                        Swal.fire({
+                            title: 'Oops!',
+                            text: response.responseJSON.message,
+                            icon: 'error',
+                            allowOutsideClick: false,
+                            allowEscapeKey: false,
+                            confirmButtonText: 'Okay',
+                            confirmButtonColor: '#0054a6',
+                        })
                     }
-                } catch (e) {
-                    console.log("Could not parse JSON response:", e);
+                    else {
+                        Swal.fire({
+                            title: 'Oops!',
+                            text: 'There was an error while processing. Please try again.',
+                            icon: 'error',
+                            allowOutsideClick: false,
+                            allowEscapeKey: false,
+                            confirmButtonText: 'Okay',
+                            confirmButtonColor: '#0054a6',
+                        })  
+                    }
+                    $('#signupFormSubmit').attr('disabled', false)
                 }
-            }
-        });
-
+            });
+        }
     });
 });